Revert "finish networking part of 24w03b port"
This reverts commit 4cbad7450838b4585aecc990b3635d7bc8ae8a31.

+++ src/main/java/net/frozenblock/lib/sound/api/instances/distance_based/RestrictedMovingFadingDistanceSwitchingSoundLoop.java
@@ -43,9 +43,9 @@ public class RestrictedMovingFadingDistanceSwitchingSoundLoop<T extends Entity> 
         this.volume = volume;
         this.pitch = pitch;
 
-        this.x = entity.getX();
-        this.y = entity.getY();
-        this.z = entity.getZ();
+        this.x = (float) entity.getX();
+        this.y = (float) entity.getY();
+        this.z = (float) entity.getZ();
         this.isFarSound = isFarSound;
         this.maxDist = maxDist;
         this.fadeDist = fadeDist;
